requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pandas==2.1.4
scikit-learn==1.3.2
lightgbm==4.5.0
xgboost==2.1.1
numpy==1.26.2
jinja2==3.1.2
Pillow==10.1.0
reportlab==4.0.8
